+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* A flutter plugin to get device information such as device IMEI number,model name,API level etc for both android & iOS.